requirements.txt
Flask
Flask-SQLAlchemy
fpdf
python-whois
requests
dnspython
pypdf
email-validator
gunicorn
waitress